+++ src/main/kotlin/cloud/glitchdev/rfu/feature/fishing/RodTimer.kt
@@ -2,6 +2,7 @@ package cloud.glitchdev.rfu.feature.fishing
 
 import cloud.glitchdev.rfu.RiccioFishingUtils.mc
 import cloud.glitchdev.rfu.config.categories.GeneralFishing
+import cloud.glitchdev.rfu.events.managers.EntityRenderEvents.registerEntityRenderEvent
 import cloud.glitchdev.rfu.events.managers.TickEvents.registerTickEvent
 import cloud.glitchdev.rfu.feature.Feature
 import cloud.glitchdev.rfu.feature.RFUFeature
@@ -14,22 +15,12 @@ import net.minecraft.world.entity.decoration.ArmorStand
 object RodTimer : Feature {
     val timerRegex = """!!!|\d\.\d""".toExactRegex()
     var timer : ArmorStand? = null
+    private var isHoldingRod = false
 
     override fun onInitialize() {
-        registerTickEvent(0, 2) { client ->
-            if(!GeneralFishing.rodTimerDisplay) return@registerTickEvent
-            val world = client.level ?: return@registerTickEvent
-
+        registerTickEvent(0, 2) { _ ->
             if(timer?.isRemoved ?: true) {
                 timer = null
-            }
-
-            if(timer == null && mc.player?.mainHandItem?.item?.descriptionId == "item.minecraft.fishing_rod") {
-                timer = world.entitiesForRendering().find { entity ->
-                    if(entity !is ArmorStand) return@find false
-                    if (!entity.hasCustomName()) return@find false
-                    return@find entity.name.toUnformattedString().matches(timerRegex)
-                } as? ArmorStand
             }
 
             RodTimerDisplay.rodTime = if(timer?.name?.string == "!!!") {
@@ -38,7 +29,27 @@ object RodTimer : Feature {
                 timer?.name?.string?.toFloatOrNull() ?: -1f
             }
 
+            isHoldingRod = mc.player?.mainHandItem?.item?.descriptionId == "item.minecraft.fishing_rod"
+
             RodTimerDisplay.updateState()
+        }
+
+        registerEntityRenderEvent { entity, event ->
+            if(!GeneralFishing.rodTimerDisplay) return@registerEntityRenderEvent
+            if(!isHoldingRod) return@registerEntityRenderEvent
+
+            if (timer != null && entity.id == timer?.id) {
+                event.cancel()
+                return@registerEntityRenderEvent
+            }
+
+            if (timer != null) return@registerEntityRenderEvent
+            if(entity !is ArmorStand) return@registerEntityRenderEvent
+            if(!entity.hasCustomName()) return@registerEntityRenderEvent
+            if(!entity.name.toUnformattedString().matches(timerRegex)) return@registerEntityRenderEvent
+
+            timer = entity
+            event.cancel()
         }
     }
 }