+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added {dh} option to rare sc catch message
 - Added option to boost polling rate when healthbar is active for more accurate tracking
 - Added message hiding options
+- Rod Timer now also hides the timer entity
 
 ### Fixes:
 - Fixed dead entities sometimes not being counted as dead (hopefully)